% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{all_sig_pal}
\alias{all_sig_pal}
\title{Color palette for mutational signatures analyzed in the study by Morrison et al.}
\format{
\subsection{\code{all_sig_pal}}{

A vector of HEX color codes for mutational signature activities.
}
}
\usage{
data(all_sig_pal)
}
\description{
Vector of named colors for use in plotting the relative activities of mutational
signatures.
}
\keyword{datasets}
