% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{carcinogens_mice_SBS.refs}
\alias{carcinogens_mice_SBS.refs}
\title{Reference single base substitution signatures found in mice exposed to 20 known or suspected carcinogens.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 96 rows and 13 columns.
}
\usage{
data(carcinogens_mice_SBS.refs)
}
\description{
Data from: Riva et al. Nat Genet (2020)
}
\keyword{datasets}
