% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{COSMIC3.0_SBS}
\alias{COSMIC3.0_SBS}
\title{Single base substitution reference signatures for COSMIC3.0}
\format{
\subsection{\code{COSMIC3.0_SBS}}{

A tibble with 96 rows and 68 columns:
\describe{
\item{Type}{Type of SBS}
\item{SBS1,...,SBS85}{Proportion of SBS in each class for each signature}
}
}
}
\source{
\url{https://cancer.sanger.ac.uk/signatures/downloads/}
}
\usage{
data(COSMIC3.0_SBS)
}
\description{
COSMIC 3.0 Single base substitutions mutational signatures
from the COSMIC website
}
\keyword{datasets}
