% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{COSMIC3.3_DBS}
\alias{COSMIC3.3_DBS}
\title{Double base reference signatures for COSMIC3.0}
\format{
\subsection{\code{COSMIC3.3_DBS}}{

A tibble with 78 rows and 12 columns:
\describe{
\item{Type}{Type of ID}
\item{DBS1,...,DBS11}{Proportion of DBS in each class for each signature}
}
}
}
\source{
\url{https://cancer.sanger.ac.uk/signatures/downloads/}
}
\usage{
data(COSMIC3.3_DBS)
}
\description{
COSMIC 3.3 Double base mutational signatures
from the COSMIC website
}
\keyword{datasets}
