% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EGFR_context_counts_ordered}
\alias{EGFR_context_counts_ordered}
\title{Counts of each context class for the main transcript of gene EGFR}
\format{
An object of class \code{table} of length 96.
}
\usage{
data(EGFR_context_counts_ordered)
}
\description{
Data from: Ensembl
}
\keyword{datasets}
