% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ESCC_countries_SBS}
\alias{ESCC_countries_SBS}
\title{Single base substitution signature attributions found in esophageal squamous cell carcinoma across countries with varying incidence.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 552 rows and 31 columns.
}
\usage{
data(ESCC_countries_SBS)
}
\description{
Data from: Moody et al. Nat Genet (2021)
}
\keyword{datasets}
