% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ESCC_countries_SBS.refs}
\alias{ESCC_countries_SBS.refs}
\title{Reference COSMIC Single base substitution signatures found in esophageal squamous cell carcinoma across countries with varying incidence.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 96 rows and 30 columns.
}
\usage{
data(ESCC_countries_SBS.refs)
}
\description{
Data from: Moody et al. Nat Genet (2021)
}
\keyword{datasets}
