% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ESCC_sig_activity}
\alias{ESCC_sig_activity}
\title{Activity of SBS, DBS, and ID mutational signatures in a global sample of esophageal squamous cell carcinoma cases.}
\format{
\subsection{\code{ESCC_sig_activity}}{

A dataframe with 552 rows and 46 columns:
\describe{
\item{Country}{Country where each sample was collected}
\item{Incidence_Level}{Incidence level (High or Low) of ESCC in the region where the sample was collected}
\item{Sample}{Unique sample ID}
\item{SBS1, ..., ID17}{Relative abundance of mutational signatures in each sample. These columns sum to 1 for each row.}
}
}
}
\source{
\url{https://doi.org/10.1038/s41588-021-00928-6}
}
\usage{
data(ESCC_sig_activity)
}
\description{
Data from: Moody, S., Senkin, S., Islam, S.M.A. et al. Mutational signatures in esophageal squamous cell carcinoma from eight countries with varying incidence. Nat Genet 53, 1553–1563 (2021).
}
\keyword{datasets}
