% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ESCC_sig_similarity}
\alias{ESCC_sig_similarity}
\title{Cosine similarity of SBS, DBS, and ID mutational signatures found in a global sample of esophageal squamous cell carcinoma cases.}
\format{
\subsection{\code{ESCC_sig_similarity}}{

A dataframe with 43 rows and 43 columns:
\describe{
\item{SBS1, ..., ID17}{Pairwise cosine similarities between mutational signatures.}
}
}
}
\source{
\url{https://doi.org/10.1038/s41588-021-00928-6}
}
\usage{
data(ESCC_sig_similarity)
}
\description{
Data from: Moody, S., Senkin, S., Islam, S.M.A. et al. Mutational signatures in esophageal squamous cell carcinoma from eight countries with varying incidence. Nat Genet 53, 1553–1563 (2021).
}
\keyword{datasets}
