% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_functions.R
\name{get_SBS96_driver_spectrum}
\alias{get_SBS96_driver_spectrum}
\title{get_SBS96_driver_spectrum}
\usage{
get_SBS96_driver_spectrum(driverlist, ref_genome)
}
\arguments{
\item{driverlist}{a table containing a list of drivers alterations, with columns chr, pos, and alt}

\item{ref_genome}{BSgenome reference genome}
}
\value{
A a named integer table (length 96) containing the SBS spectrum of the transcript
}
\description{
get_SBS96_driver_spectrum
}
\examples{
data(TP53_drivers_intogen_LUAD, package = "sigvar")
ref_genome <- "BSgenome.Hsapiens.UCSC.hg38"
library(ref_genome, character.only = TRUE)
TP53_LUAD.driver.spectrum <- get_SBS96_driver_spectrum(TP53_drivers_intogen_LUAD,ref_genome)
print(TP53_LUAD.driver.spectrum)
}
