% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Intogen_EGFR_LUAD}
\alias{Intogen_EGFR_LUAD}
\title{Driver mutations from the intogen website for gene EGFR in lung adenocarcinomas}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 25 rows and 13 columns.
}
\usage{
data(Intogen_EGFR_LUAD)
}
\description{
Data from: Intogen database
}
\keyword{datasets}
