% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MESOMICS_CN_SBS}
\alias{MESOMICS_CN_SBS}
\title{Copy Number and Single Base Substitution mutational signatures for 120 malignant pleural mesothelioma from the MESOMICS cohort}
\format{
\subsection{\code{MESOMICS_CN_SBS}}{

A tibble with 120 rows and 41 columns:
\describe{
\item{ID_MESOMICS}{Identifier of the patient}
\item{Sample}{Identifier of the tumor sample}
\item{Sex}{Patient sex}
\item{Age}{Patient age at diagnosis}
\item{Smoking, Professional.Asbestos, ...}{Patient exposure status}
\item{Survival.Censor, Survival.Time}{Survival status (dead or alive) and time in months}
\item{Asbestos.Exposure.probability, Asbestos.Exposure.frequency, ...}{Quantitative asbestos exposure variables}
\item{Type}{Tumor histological type (WHO)}
\item{CNB}{Copy number burden}
\item{CN1,...,CN19}{Copy number signature attributions}
\item{SBS1,...,SBS40}{Single Base Substitution signature attributions}
}
}
}
\source{
\url{https://static-content.springer.com/esm/art\%3A10.1038\%2Fs41588-023-01321-1/MediaObjects/41588_2023_1321_MOESM4_ESM.xlsx}
}
\usage{
data(MESOMICS_CN_SBS)
}
\description{
Copy Number and Single Base Substitution mutational signature attributions
from Mangiante et al. Nat Genet 55, 607–618 (2023). https://doi.org/10.1038/s41588-023-01321-1
}
\keyword{datasets}
