% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PCAWG_SigProfiler_COSMIC_DBS}
\alias{PCAWG_SigProfiler_COSMIC_DBS}
\title{Double Base Substitution COSMIC mutational signatures for 2780 tumors}
\format{
\subsection{\code{PCAWG_SigProfiler_COSMIC_DBS}}{

A tibble with 2,670 rows and 14 columns:
\describe{
\item{Cancer Types}{Type of cancer}
\item{Sample Names}{Identifier of the tumor}
\item{DBS1, DBS2}{Number of mutations for each signature}
...
}
}
}
\source{
\url{https://dcc.icgc.org/releases/PCAWG/mutational_signatures/Signatures_in_Samples/SP_Signatures_in_Samples}
}
\usage{
data(PCAWG_SigProfiler_COSMIC_DBS)
}
\description{
Double Base Substitution mutational signature attributions to COSMIC signatures
from the Pan Cancer Whole-Genome consortium (PCAWG). Oct 19 2019 version
}
\keyword{datasets}
