% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_vis_functions.R
\name{plot_SBS_spectrum}
\alias{plot_SBS_spectrum}
\title{Plot an SBS mutational spectrum}
\usage{
plot_SBS_spectrum(SBS_table)
}
\arguments{
\item{SBS_table}{A matrix with rows corresponding to the 96 single-base substitutions and columns corresponding to distinct mutational spectra you wish to plot.}
}
\value{
A ggplot2 bar chart depicting SBS mutational spectra
}
\description{
Plot an SBS mutational spectrum
}
\examples{
data(COSMIC3.3.1_SBS, package = "sigvar")
SBS_table <- dplyr::select(COSMIC3.3.1_SBS, SBS1, SBS5)
plot_SBS_spectrum(SBS_table)

}
