% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_vis_functions.R
\name{plot_signature_prop}
\alias{plot_signature_prop}
\title{Plot the signature proportion for a set of samples}
\usage{
plot_signature_prop(Q)
}
\arguments{
\item{Q}{A dataframe, matrix, or array representing mutational signature
relative contributions.
Each row represents a sample.}
}
\description{
Plot the signature proportion for a set of samples
}
