% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rcc_sig_activity}
\alias{rcc_sig_activity}
\title{Activity of SBS, DBS, and ID mutational signatures in a global sample of renal cell carcinoma cases.}
\format{
\subsection{\code{rcc_sig_activity}}{

A dataframe with 962 rows and 31 columns:
\describe{
\item{Sample}{Unique sample ID}
\item{Country}{Country where each sample was collected}
\item{ASR}{Age-standardized rate (ASR) of RCC incidence in each country}
\item{SBS1, ..., ID_C}{Relative abundance of mutational signatures in each sample. These columns sum to 1 for each row.}
}
}
}
\source{
\url{https://doi.org/10.1101/2023.06.20.23291538}
}
\usage{
data(rcc_sig_activity)
}
\description{
Data from: Senkin et al. (2023). Geographic variation of mutagenic exposures in kidney cancer genomes. MedRxiv, 14, 2023.06.20.23291538. https://doi.org/10.1101/2023.06.20.23291538
}
\keyword{datasets}
