% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rcc_sim}
\alias{rcc_sim}
\title{Activity of SBS, DBS, and ID mutational signatures in a global sample of renal cell carcinoma cases.}
\format{
\subsection{\code{rcc_sim}}{

A dataframe with 28 rows and 28 columns:
\describe{
\item{SBS1, ..., ID_C}{Cosine similarity between mutational signatures.}
}
}
}
\source{
\url{https://doi.org/10.1101/2023.06.20.23291538}
}
\usage{
data(rcc_sim)
}
\description{
Data from: Senkin et al. (2023). Geographic variation of mutagenic exposures in kidney cancer genomes. MedRxiv, 14, 2023.06.20.23291538. https://doi.org/10.1101/2023.06.20.23291538
}
\keyword{datasets}
