% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sbs_palette}
\alias{sbs_palette}
\title{Single Base Substitution COSMIC mutational signatures for 2780 tumors}
\format{
\subsection{\code{sbs_palette}}{

A vector of RGB color codes for COSMIC SBS signatures.
}
}
\source{
\url{https://dcc.icgc.org/releases/PCAWG/mutational_signatures/Signatures_in_Samples/SP_Signatures_in_Samples}
}
\usage{
data(sbs_palette)
}
\description{
Single Base Substitution mutational signature attributions to COSMIC signatures
from the Pan Cancer Whole-Genome consortium (PCAWG). Oct 19 2019 version
}
\keyword{datasets}
