% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Sherlock_LCINS.metadata}
\alias{Sherlock_LCINS.metadata}
\title{Metadata accompanying mutational signatures for lung cancer in never smokers.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 232 rows and 24 columns.
}
\usage{
data(Sherlock_LCINS.metadata)
}
\description{
Data from: Zhang et al. Nat Genet (2021)
}
\keyword{datasets}
