% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Sherlock_LCINS_SBS.refs}
\alias{Sherlock_LCINS_SBS.refs}
\title{Reference COSMIC Single base substitution signatures found in Lung cancer in never smokers.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 96 rows and 16 columns.
}
\usage{
data(Sherlock_LCINS_SBS.refs)
}
\description{
Data from: Zhang et al. Nat Genet (2021)
}
\keyword{datasets}
