% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_functions.R
\name{sigFAVA}
\alias{sigFAVA}
\title{sigFAVA
Main function for FAVA. Can test the statistical association between the variability of a Q matrix and a covariable
using a moving average with a Gaussian kernel}
\usage{
sigFAVA()
}
\arguments{
\item{Q}{a matrix of mutational signatures in multiple individuals}

\item{w}{the bandwidth of the Gaussian kernel}
}
\value{
Test statistics
}
\description{
sigFAVA
Main function for FAVA. Can test the statistical association between the variability of a Q matrix and a covariable
using a moving average with a Gaussian kernel
}
\examples{
#Import Q matrix of carcinogens in mice from Riva et al. Nature Genetics 2020
Qmice = as.data.frame(mutsig_carcinogens_mice_SBS[,c(2:12,25)])
Qmice[Qmice<0.1] = 0
plot_dots(Qmice, group = "chemical" )
#test
test_covariable(Qlist[[1]],1,1)
}
