% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{smoker_sigs_chen}
\alias{smoker_sigs_chen}
\title{Activity of Smoking, APOBEC, and Ageing mutational signatures in a sample of East Asian lung adenocarcinoma patients.}
\format{
\subsection{\code{smoker_sigs_chen}}{

A dataframe with 88 rows and 20 columns:
\describe{
\item{Patient.ID}{Unique identifier for each patient}
\item{Smoker}{Smoking status for each patient: "Smoker" or "Non-smoker"}
\item{Cohort, Stage, Age, Gender}{Patient metadata}
\item{Smoking}{Activity of smoking mutational signature}
\item{APOBEC}{Activity of APOBEC mutational signature}
\item{Ageing}{Activity of ageing mutational signature}
}
}
}
\source{
\url{https://doi.org/10.1038/s41588-019-0569-6}
}
\usage{
data(smoker_sigs_chen)
}
\description{
Data from: Chen, J., Yang, H., Teo, A.S.M. et al. Genomic landscape of lung adenocarcinoma in East Asians. Nat Genet 52, 177–186 (2020).
}
\keyword{datasets}
