% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TP53_drivers_intogen_SBS96}
\alias{TP53_drivers_intogen_SBS96}
\title{SBS96 spectrum from the intogen website for gene TP53 in lung adenocarcinomas}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 96 rows and 4 columns.
}
\usage{
data(TP53_drivers_intogen_SBS96)
}
\description{
Data from: Intogen database
}
\keyword{datasets}
