% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zhang_sig_activity}
\alias{zhang_sig_activity}
\title{Activity of SBS mutational signatures in samples of lung cancer in never smokers}
\format{
\subsection{\code{zhang_sig_activity}}{

A dataframe with 39 rows and 213 columns:
\describe{
\item{Subject}{Unique ID for each subject}
...
\item{passive_smoking}{Subject smoking status ("Non-smoker" or "Passive smoker")}
...
\item{SBS1, ..., SBS40}{Relative abundance of mutational signatures in each sample. These columns sum to 1 for each row.}
}
}
}
\source{
\url{https://doi.org/10.1038/s41588-021-00920-0}
}
\usage{
data(zhang_sig_activity)
}
\description{
Data from: Zhang et al. (2021). Genomic and evolutionary classification of lung cancer in never smokers. Nature Genetics, 53(9), 1348–1359. https://doi.org/10.1038/s41588-021-00920-0
}
\keyword{datasets}
