% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zhang_sig_similarity}
\alias{zhang_sig_similarity}
\title{Cosine similarity of SBS mutational signatures found in samples of lung cancer in never smokers}
\format{
\subsection{\code{zhang_sig_similarity}}{

A dataframe with 14 rows and 14 columns:
\describe{
\item{SBS1, ..., SBS40}{Pairwise cosine similarities between mutational signatures.}
}
}
}
\source{
\url{https://doi.org/10.1038/s41588-021-00920-0}
}
\usage{
data(zhang_sig_similarity)
}
\description{
Data from: Zhang et al. (2021). Genomic and evolutionary classification of lung cancer in never smokers. Nature Genetics, 53(9), 1348–1359. https://doi.org/10.1038/s41588-021-00920-0
}
\keyword{datasets}
