% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{carcinogens_mice_SBS}
\alias{carcinogens_mice_SBS}
\title{Single base substitution signature attributions found in mice exposed to 20 known or suspected carcinogens.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 181 rows and 47 columns.
}
\usage{
carcinogens_mice_SBS
}
\description{
Data from: Riva et al. Nat Genet (2020)
}
\keyword{datasets}
