% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{COSMIC3.0_ID}
\alias{COSMIC3.0_ID}
\title{Indel reference signatures for COSMIC3.0}
\format{
\subsection{\code{COSMIC3.0_ID}}{

A tibble with 83 rows and 18 columns:
\describe{
\item{Type}{Type of ID}
\item{ID1,...,ID17}{Proportion of ID in each class for each signature}
}
}
}
\source{
\url{https://cancer.sanger.ac.uk/signatures/downloads/}
}
\usage{
COSMIC3.0_ID
}
\description{
COSMIC 3.0 Indel mutational signatures
from the COSMIC website
}
\keyword{datasets}
