% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{COSMIC3.3_CN}
\alias{COSMIC3.3_CN}
\title{Copy Number reference signatures for COSMIC3.3}
\format{
\subsection{\code{COSMIC3.3_CN}}{

A tibble with 48 rows and 8 columns:
\describe{
\item{MutationsType}{Type of segment}
\item{CN1,...,CN19}{Proportion of Copy number segments in each category}
}
}
}
\source{
\url{https://cancer.sanger.ac.uk/signatures/downloads/}
}
\usage{
COSMIC3.3_CN
}
\description{
COSMIC 3.3 Copy Number mutational signatures
from the COSMIC website
}
\keyword{datasets}
