% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_functions.R
\name{cossim}
\alias{cossim}
\title{Compute the cosine similarity matrix for a set of mutational signatures.}
\usage{
cossim(ref_sigs)
}
\arguments{
\item{ref_sigs}{A matrix with K columns containing non-negative entries that sum to 1. Each column represents a signature (e.g., SBS1, SBS2), each row represents a mutation type (e.g., \code{A[C>A]A}), and each entry represents the abundance of that type in the signature.}
}
\value{
A K x K similarity matrix with diagonal elements equal to 1 and off-diagonal elements between 0 and 1. Entry \code{S[j,k]} is the similarity between signature \code{j} and signature \code{k}, equaling 1 if the categories are to be treated as identical and equaling 0 if they are to be treated as totally dissimilar. The default value is \code{S = diag(K)}.
}
\description{
Compute the cosine similarity matrix for a set of mutational signatures.
}
\examples{
# Compute the cosine similarity matrix for the lung cancer
# in never smoker datasets from Zhang et al. 2021

 cossim(ref_sigs = as.matrix(Sherlock_LCINS_SBS.refs[,1:14]))

}
