% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_functions.R
\name{get_SBS96_driver_spectrum}
\alias{get_SBS96_driver_spectrum}
\title{get_SBS96_spectrum_driver}
\usage{
get_SBS96_driver_spectrum(driverlist)
}
\arguments{
\item{driverlist}{a table containing a list of drivers alterations, with columns chr, pos, and alt}
}
\value{
A vector of size 96 containing the SBS spectrum of the transcript
}
\description{
get_SBS96_spectrum_driver
}
\examples{
TP53_LUAD.driver.spectrum = get_SBS96_driver_spectrum(TP53_drivers_intogen_LUAD)
print(TP53_LUAD.driver.spectrum)
}
