% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_functions.R
\name{get_SBS96_spectrum}
\alias{get_SBS96_spectrum}
\title{get_SBS96_spectrum}
\usage{
get_SBS96_spectrum(transcript = "ENST00000269305.9", organism = "Homo sapiens")
}
\arguments{
\item{transcript}{the ensembl ID of the transcript}

\item{organism}{the name of the organism associated with the transcript (species)}
}
\value{
A vector of size 96 containing the SBS spectrum of the transcript
}
\description{
get_SBS96_spectrum
}
\examples{
#Run on
spectrum = get_SBS96_spectrum(transcript = "ENST00000269305.9")
print(spectrum)
}
