% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_functions.R
\name{import_SigProfiler}
\alias{import_SigProfiler}
\title{import_SigProfiler}
\usage{
import_SigProfiler(folder = ".")
}
\arguments{
\item{folder}{a folder containing SigProfilerExtractor results}
}
\value{
A list of tibbles containing the SigProfiler activity matrices
}
\description{
import_SigProfiler
}
