% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mutsig_carcinogens_mice_SBS.refs}
\alias{mutsig_carcinogens_mice_SBS.refs}
\title{Definitions of murine single base substitution mutational signatures}
\format{
\subsection{\code{mutsig_carcinogens_mice_SBS.refs}}{

A tibble with 96 rows and 13 columns:
\describe{
\item{Type}{Substitution type (e.g., C>A, T>G)}
\item{Subtype}{Substitution type with 5' and 3' context (e.g., ACA)}
\item{mSBS5, ..., mSBS_N3}{Proportion of mutations in each class for each signature}
}
}
}
\usage{
mutsig_carcinogens_mice_SBS.refs
}
\description{
Single Base Substitution mutational signature definitions
from Riva et al. Nat Genet 52, 1189–1197 (2020). https://doi.org/10.1038/s41588-020-0692-4
}
\keyword{datasets}
