% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{radiation_sigs_morton}
\alias{radiation_sigs_morton}
\title{Single base substitution signature attributions found in papillary thyroid carcinoma from Chernobyl incident survivors.}
\format{
An object of class \code{data.frame} with 356 rows and 16 columns.
}
\usage{
radiation_sigs_morton
}
\description{
Data from: Morton et al. Science (2021)
}
\keyword{datasets}
