% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{radiation_sigs_morton_cossim}
\alias{radiation_sigs_morton_cossim}
\title{Cosine similarity matrix of Single base substitution signatures found in papillary thyroid carcinoma from Chernobyl incident survivors.}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 13 rows and 13 columns.
}
\usage{
radiation_sigs_morton_cossim
}
\description{
Data from: Morton et al. Science (2021)
}
\keyword{datasets}
