% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Sherlock_LCINS_SBS}
\alias{Sherlock_LCINS_SBS}
\title{Single base substitution signature attributions found in Lung cancer in never smokers.}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 232 rows and 15 columns.
}
\usage{
Sherlock_LCINS_SBS
}
\description{
Data from: Zhang et al. Nat Genet (2021)
}
\keyword{datasets}
