% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{sigboot}
\alias{sigboot}
\title{Statistically compare the within-sample diversity and across-sample heterogeneity of the mutational signature activity in two or more groups of tumor samples.}
\usage{
sigboot(
  sig_activity,
  n_replicates,
  K = ncol(sig_activity) - length(group),
  group,
  S = NULL,
  normalized = FALSE,
  seed = NULL,
  save_replicates = FALSE,
  alternative = "two.sided"
)
}
\arguments{
\item{sig_activity}{A matrix or data frame with rows containing non-negative entries that sum to 1. Each row represents a sample, each column represents a mutational signature, and each entry represents the abundance of that signature in the sample. If \code{sig_activity} contains any metadata, it must be on the left-hand side of the matrix, the right \code{K} entries of each row must sum to 1, and \code{K} must be specified. Otherwise, all entries of each row must sum to 1.}

\item{n_replicates}{The number of bootstrap replicate matrices to generate for each provided relative abundance matrix.}

\item{K}{Optional; an integer specifying the number of mutational signatures included in \code{sig_activity}. Default is \code{K=ncol(sig_activity)-length(group)}.}

\item{group}{A string (or vector of strings) specifying the name(s) of the column(s) that describes which group(s) each sample belongs to.}

\item{S}{Optional; a K x K similarity matrix with diagonal elements equal to 1 and off-diagonal elements between 0 and 1. Entry \code{S[j,k]} is the similarity between signature \code{j} and signature \code{k}, equaling 1 if the categories are to be treated as identical and equaling 0 if they are to be treated as totally dissimilar. The default value is \code{S = diag(K)}.}

\item{normalized}{Optional; should the across-sample variability (Fst) be normalized by its upper bound conditional on the mean activity of the most abundant signature be used as the measure of across-sample variability? Default is \code{normalized = FALSE}; use \code{normalized = TRUE} to compute normalized Fst. Fst can only be normalized if it is not weighted.}

\item{seed}{Optional; an integer to be used as a random seed for the simulations.}

\item{save_replicates}{Optional; should all of the bootstrap replicate matrices be included in the output? Default is \code{save_replicates = FALSE};  \code{save_replicates = FALSE} savea memory when analyzing large datasets.}

\item{alternative}{Optional; do you want to do a one- or two.sided test? Default is \code{alternative = "two.sided"}. If you wish to do a one-sided test, specify either \code{alternative = "less"} or \code{alternative = "greater"}.}
}
\value{
A named list containing the following entries:
\itemize{
\item \code{p_values}: The probability of observing the observed difference in variability between each pair of groups if there were no difference between groups. Computed as the fraction of bootstrap differences greater than or equal to the observed difference. Depends on what \code{alternative} is specified ("greater", "lesser", or "two.sided").
\item \code{bootstrap_distribution_plot}: The distribution of bootstrap replicate differences in each variability value. The observed differences are shown in red. The further the red points are from 0, the more significant the statistical difference between groups.
\item \code{observed_stats}: The observed diversity statistics for the groups.
\item \code{bootstrap_stats}: The bootstrap replicate diversity statistics for the groups.
\item \code{bootstrap_replicates}: The bootstrap replicate matrices, reported only if  \code{save_replicates = TRUE}.
}
\description{
\code{sigboot} uses bootstrapping to evaluate differences in within-sample diversity and across-sample heterogeneity of mutational signature activity  between pairs of groups of tumor samples. \code{sigboot} takes the same options as \code{sigvar}, so, as with \code{sigvar}, you can separately analyze multiple populations or groups of samples (specify \code{group}), and account for cosine similarity among signatures (specify \code{S}). \code{sigboot} follows the bootstrapping procedure defined by Efron and Tibshirani (1993). Details on the bootstrapping procedure are available in the Methods section of the accompanying paper.
}
\examples{
# Estimate the uncertainty in the across-sample and mean within-sample variability of
mutational signatures in ESCC samples grouped by country
# We provide a cosine similarity matrix in order to account for cosine similarity among signatures
smoker_boot = sigboot(sig_activity = smoker_sigs_chen, K = 3, n_replicates = 500,
                    group = "Smoker", S = smoker_sigs_chen_cossim,
                    seed = 1)

# We find that there is a strong, significant difference between smokers and non-smokers
# in their mean within-sample diversity, but that the difference in across-sample heterogeneity
# is less strong
smoker_boot$P_values

# We can visualize this result by looking at the observed differences between
# smokers and non-smokers and the bootstrap replicates, which assume no difference
# between these two groups
smoker_boot$bootstrap_distribution_plot


}
