% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_functions.R
\name{sigFAVA}
\alias{sigFAVA}
\title{#' sigboot
#' Estimate uncertainty in the within- and mean across-sample variability of mutational signature activity in one or multiple populations of samples.
#'
#' Generate bootstrap replicates of one or more tables of mutational signature activity. Compute across-sample and mean within-sample variability of each replicate table. Visualize the resulting distributions of variability values and compare populations with statistical tests. \code{sigboot} takes the same options as \code{sigvar}, so, as with \code{sigvar}, you can separately analyze multiple populations or groups of samples (specify \code{group}), account for cosine similarity among signatures (specify \code{S}), and incorporate non-uniform weightings of samples (specify \code{w} or \code{time}).
#'
#' @param sig_activity A matrix or data frame with rows containing non-negative entries that sum to 1. Each row represents a sample, each column represents a mutational signature, and each entry represents the abundance of that signature in the sample. If \code{sig_activity} contains any metadata, it must be on the left-hand side of the matrix, the right \code{K} entries of each row must sum to 1, and \code{K} must be specified. Otherwise, all entries of each row must sum to 1.
#' @param n_replicates The number of bootstrap replicate matrices to generate for each provided relative abundance matrix.
#' @param K Optional; an integer specifying the number of mutational signatures included in \code{sig_activity}. Default is \code{K=ncol(sig_activity)}.
#' @param group Optional; a string (or vector of strings) specifying the name(s) of the column(s) that describes which group(s) each sample belongs to. Use if \code{sig_activity} is a matrix containing multiple groups of samples you wish to compare.
#' @param S Optional; a K x K similarity matrix with diagonal elements equal to 1 and off-diagonal elements between 0 and 1. Entry \code{S[j,k]} is the similarity between signature \code{j} and signature \code{k}, equaling 1 if the categories are to be treated as identical and equaling 0 if they are to be treated as totally dissimilar. The default value is \code{S = diag(K)}.
#' @param w Optional; a vector of length \code{nrow(sig_activity)} with non-negative entries that sum to 1. Entry \code{w[i]} represents the weight placed on row \code{i} in the computation of the mean abundance of each category across rows. The default value is \code{w = rep(1/nrow(sig_activity), nrow(sig_activity))}.
#' @param time Optional; a string specifying the name of the column that describes the sampling time for each row. Include if you wish to weight the variability measures by the distance between adjacent samples in the matrix.
#' @param normalized Optional; should Fst normalized by its upper bound conditional on the mean activity of the most abundant signature be used as the measure of across-sample variability? Default is \code{normalized = FALSE}; use \code{normalized = TRUE} to compute normalized FAVA. FAVA can only be normalized if it is not weighted.
#' @param save_replicates Optional; should all of the bootstrap replicate matrices be included in the output? Default is \code{save_replicates = FALSE};  \code{save_replicates = FALSE} savea memory when analyzing large datasets.
#'
#' @return A named list containing the following entries:
#' \itemize{
#' \item \code{bootstrap_replicates}: If \code{save_replicates = TRUE}, a list of lists. Each element is named for a group provided in \code{sig_activity} and contains a list of \code{n_replicates} bootstrap replicates of the provided matrix. E.g., if \code{n_replicates = 100} and the first group in \code{sig_activity} is named \code{A}, then the first element of \code{bootstrap_replicates} is itself a list of 100 matrices, each representing a bootstrap replicate of matrix A.
#' \item \code{statistics}: A dataframe containing \code{Across_sample_variability} and \code{Mean_within_sample_variability} computed for each bootstrap replicate matrix in \code{bootstrap_replicates}. If \code{group} is specified, the preceding columns indicate which group, or combination of groups, the row corresponds to.
#' \item \code{plot_variability}: A ggplot2 scatter plot depicting the bootstrap distributions of mean within-sample variability (x-axis) and across-sample variability (y-axis) for each group in \code{sig_activity}. Ellipses provide 95\% confidence regions.
#' \item \code{plot_boxplot}: A ggplot2 boxplot separately depicting the bootstrap distribution of each type of variability for each group.
#' \item \code{test_kruskal_wallis_across}: Results of a Kruskal-Wallis test performed on the bootstrap distributions of across-sample variability. This test is a non-parametric statistical test of whether all provided bootstrap distributions are identically distributed.
#' \item \code{test_pairwise_wilcox_across}: Results of a Wilcoxon rank-sum test performed on the bootstrap distributions of across-sample variability. This test is a non-parameteric statistical test of whether \emph{each pairwise combination} of provided bootstrap distributions is identically distributed. The result is a matrix of p-values whose entries correspond to each pair of relative abundance matrices.
#' #' \item \code{test_kruskal_wallis_within}: Results of a Kruskal-Wallis test performed on the bootstrap distributions of mean within-sample variability. This test is a non-parametric statistical test of whether all provided bootstrap distributions are identically distributed.
#' \item \code{test_pairwise_wilcox_within}: Results of a Wilcoxon rank-sum test performed on the bootstrap distributions of mean within-sample variability. This test is a non-parameteric statistical test of whether \emph{each pairwise combination} of provided bootstrap distributions is identically distributed. The result is a matrix of p-values whose entries correspond to each pair of relative abundance matrices.
#' @export
#'
#' @examples
#' # Estimate the uncertainty in the across-sample and mean within-sample variability of
#' mutational signatures in ESCC samples grouped by country
#' # We provide a cosine similarity matrix in order to account for cosine similarity among signatures
#' smoker_boot = sigboot(sig_activity = smoker_sigs_chen, K = 3, n_replicates = 500,
#'                     group = "Smoker", S = smoker_sigs_chen_cossim,
#'                     seed = 1)
#'
#' # The statistics entry of the output gives bootstrap replicates of the two
#' # variability measures for each group
#' head(smoker_boot$statistics)
#'
#' # sigboot visualizes these bootstrap replicates in two types of plots:
#' smoker_boot$plot_variability
#' smoker_boot$plot_boxplot
#'
#' # sigboot also computes statistics comparing the two types of variability for each group:
#' smoker_boot$test_pairwise_wilcox_across
#' smoker_boot$test_pairwise_wilcox_within
#'
#' @import FAVA
sigboot <- function(sig_activity, n_replicates,
K = NULL,
group = NULL,
seed = NULL,
S = NULL,
w = NULL,
time = NULL,
normalized = FALSE,
save_replicates = FALSE){}
\usage{
sigFAVA()
}
\arguments{
\item{Q}{a matrix of mutational signatures in multiple individuals}

\item{w}{the bandwidth of the Gaussian kernel}
}
\value{
Test statistics
}
\description{
boot_out = bootstrap_fava(sig_activity,
K = K, S = S,
n_replicates = n_replicates, seed = seed, group = group,
time = time, w = w, normalized = normalized, save_replicates = save_replicates)
}
\details{

}
\examples{
#Import Q matrix of carcinogens in mice from Riva et al. Nature Genetics 2020
Qmice = as.data.frame(mutsig_carcinogens_mice_SBS[,c(2:12,25)])
Qmice[Qmice<0.1] = 0
plot_dots(Qmice, group = "chemical" )
#test
test_covariable(Qlist[[1]],1,1)
}
