% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigvar-package.R
\docType{package}
\name{sigvar-package}
\alias{sigvar-package}
\title{sigvar: Quantify and visualize variability of mutational signatures within and across samples}
\description{
This package allows users to import mutational signature attribution (a.k.a. exposure) matrices and compute, visualize, and test their variabilities within and across samples.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MaikeMorrison/sigvar}
  \item Report bugs at \url{https://github.com/MaikeMorrison/sigvar/issues}
}

}
\author{
\strong{Maintainer}: Nicolas Alcala \email{alcalan@iarc.who.int} (\href{https://orcid.org/0000-0002-5961-5064}{ORCID})

Authors:
\itemize{
  \item Nicolas Alcala \email{alcalan@iarc.who.int} (\href{https://orcid.org/0000-0002-5961-5064}{ORCID})
  \item Maike Morrison \email{maike.morrison@gmail.com} (\href{https://orcid.org/0000-0003-0430-1401}{ORCID})
}

Other contributors:
\itemize{
  \item Worldwide Cancer Research (Grant 24-0106) [funder]
  \item French Ligue Nationale Contre le Cancer [funder]
}

}
\keyword{internal}
