% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{smoker_sigs_chen_cossim}
\alias{smoker_sigs_chen_cossim}
\title{Cosine similarity matrix of Single base substitution signatures found in lung adenocarcinomas in Asia.}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 3 rows and 3 columns.
}
\usage{
smoker_sigs_chen_cossim
}
\description{
Data from: Chen et al. Nat Genet (2020)
}
\keyword{datasets}
