% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TP53_context_counts_ordered}
\alias{TP53_context_counts_ordered}
\title{Counts of each context class for the main transcript of gene TP53}
\format{
An object of class \code{table} of length 96.
}
\usage{
TP53_context_counts_ordered
}
\description{
Data from: Ensembl
}
\keyword{datasets}
