% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TP53_drivers_intogen_LUAD}
\alias{TP53_drivers_intogen_LUAD}
\title{Driver mutations from the intogen website for gene TP53 in lung adenocarcinomas}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 202 rows and 10 columns.
}
\usage{
TP53_drivers_intogen_LUAD
}
\description{
Data from: Intogen database
}
\keyword{datasets}
